Made no from int to float

+++ app.py
@@ -13,8 +13,8 @@ btn = st.button('Calculate')
 if btn:
     if num1 and num2:
         try:
-            num1 = int(num1)
-            num2 = int(num2)
+            num1 = float(num1)
+            num2 = float(num2)
             st.markdown(f"## {num1} - {num2} = {float(num1) - float(num2)}")
         except ZeroDivisionError:
             st.error("Do not divide by zero !!!")
